@Rifur13 2024-08-16
Here are the numbers compared the previous kernel. Trying different values of **num_wraps**,  **num_wraps** and block sizes can be helpful to further increase performance for your hardware.

Note that the **improvement** is the relative speedup, and not just the percentage increase.


| Batch Size | Num Q Heads | Num KV Heads | Q Seq Len | KV Seq Len | Head Dim | DType | Kernel | Baseline | Improvement |
|---|---|---|---|---|---|---|---|---|---|
| 16 | 24 | 24 | 328 | 328 | 32 | bfloat16 | 411.552us +/- 6.66 | 399.462us +/- 3.72 | -3.03% |
| 16 | 24 | 24 | 328 | 328 | 64 | bfloat16 | 642.733us +/- 1.68 | 666.949us +/- 2.04 | +3.63% |
| 16 | 24 | 24 | 328 | 328 | 128 | bfloat16 | 1214.854us +/- 4.02 | 2955.590us +/- 3.62 | +58.90% |
| 4 | 24 | 24 | 4096 | 4096 | 32 | bfloat16 | 9102.094us +/- 12.43 | 9859.752us +/- 7.32 | +7.68% |
| 4 | 24 | 24 | 4096 | 4096 | 64 | bfloat16 | 14404.937us +/- 59.16 | 14744.581us +/- 18.49 | +2.30% |
| 4 | 24 | 24 | 4096 | 4096 | 128 | bfloat16 | 21383.099us +/- 110.85 | 64695.606us +/- 65.18 | +66.95% |
| 1 | 2 | 2 | 32768 | 32768 | 32 | bfloat16 | 11271.190us +/- 938.29 | 411253.577us +/- 486.08 | +97.26% |
| 1 | 2 | 2 | 32768 | 32768 | 64 | bfloat16 | 17974.650us +/- 1989.63 | 672074.326us +/- 520.27 | +97.33% |
| 1 | 2 | 2 | 32768 | 32768 | 128 | bfloat16 | 30878.254us +/- 3374.88 | 2278354.646us +/- 4696.14 | +98.64% |